Builder pattern added to project
This one is pretty big update. Finally everything is working both in
tests and inside the app. The builder was created with lombok assistance
- by using builder annotation. I also had to add blank constructor and
mark it with tolerate annotation because of nature of the project.

There were also some small fixes and code cleaning.

Altough it was all tested im not 100% sure there won't be any problems
with it. I'm waiting for suggestions.